refactor: create opponents module
